fix some issues with code
Also move the code around a little bit
